Fixed the getters that return pointers instead of copying values

+++ src/xdpd/drivers/gnu_linux/src/io/packet_classifiers/c_pktclassifier/headers/cpc_vlan.h
@@ -39,7 +39,7 @@ void set_vlan_pcp(void* hdr, uint8_t pcp){
 }
 
 inline static
-uint16_t* get_vlan_pcp(void* hdr){
+uint8_t* get_vlan_pcp(void* hdr){
 	return &((cpc_vlan_hdr_t*)hdr)->byte0;
 }
